google-fit-data/requirements.txt
google-api-python-client>=2.0.0
httplib2>=0.22.0
oauth2client>=4.1.3
python-dateutil>=2.8.0
pyyaml>=6.0
pyasn1>=0.5.0
pyasn1-modules>=0.3.0
rsa>=4.9
six>=1.16.0
uritemplate>=4.1.0
